requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=1.0.1
tqdm>=4.66.0
